+++ frontend/src/hooks/useProxy.jsx
@@ -1,0 +1,4 @@
+export const useProxy = (source) => {
+  const split = source.split('imageserver-2022/');
+  return 'https://lucascranach.org/data-proxy/image.php?subpath=/' + split[1];
+};
